Only automatically run queries
We don't want to automatically run anything that's not a query.
In this case, I've turned off both mutations and subscriptions.
This is slightly overkill since I don't think subscriptions can
issue "write"-type operations but just to be safe we'll start
with them off and later when we have subscriptions, we can
evaluate turning them back on.